// drr/src/UI/src/model/increasedOrTransferred.ts
/**
 * Generated by orval v7.2.0 🍺
 * Do not edit manually.
 * DRR API
 * OpenAPI spec version: 1.0.0
 */

export type IncreasedOrTransferred = typeof IncreasedOrTransferred[keyof typeof IncreasedOrTransferred];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const IncreasedOrTransferred = {
  Increased: 'Increased',
  Transferred: 'Transferred',
} as const;